Tweeks and fixes for the Atom feed
1. In the FeedConfiguration record, remove an extraneous forward slash in the feedRoot URL.

2. Also, add a more detailed feedDescription.

3. Hakyll's template for Atom does not actually use feedDescription,
even though its documentation advises it is a mandatory value to provide.
Studying the RFC's for RSS and Atom, we see that the description tag for RSS
has been replaced by the subtitle tag for Atom.
Therefore, we expose the default atom template,
add a subtitle tag and have it filled by feedDescription.

4. The default atom-item template uses the summary tag for containing content,
but if we look closely at the Atom feed, we see that complete blog contents are being used.
Therefore, we also expose the atom-item template and change the summary tag to content tag.

There were two errors in commit 74ac828 that is also worth fixing.

5. A link to the Atom feed was added to the header template,
but this template is used in all index.html pages, not just the main one,
which means that all index.html pages should be included in the commit.

6. The link to the Atom feed was not absolute,
which means that only the main index.html would point to a working feed.
The Atom feed link is made absolute by using the feedRoot value from the feedConfiguration record.
The feedRoot value is added to the defaultContext that is used to render the header template.